Refactor: shuffle push & worker concerns out of core
