+++ types/src/core/push/index.d.ts
@@ -1,11 +1,11 @@
 import { Context } from '..';
 export declare type PushSubscriptionState = 'subscribed' | 'unsubscribed' | 'blocked';
-export declare enum MessageType {
-    PUSH = 1
-}
 export declare enum TokenType {
     W3C = 3,
     SAFARI = 4
+}
+export declare enum MessageType {
+    PUSH = 1
 }
 export interface PushOpsManager {
     requestNotificationPermission(ctx: Context): Promise<NotificationPermission>;
