+++ types/src/worker/utils.d.ts
@@ -1,3 +1,3 @@
-import { WorkerMessage } from '../core';
+import { WorkerMessage } from './messaging';
 export declare function postEventToClient(event: WorkerMessage, client: Client): void;
 export declare function broadcastFromWorker(event: WorkerMessage): Promise<void>;
